optimize read
